+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.28.3 is now available from:
 
   <https://download.bitcoinabc.org/0.28.3/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features:
+ - Add 2 new RPCs, `invalidateavalancheproof` and `reconsideravalancheproof` to
+   manually change the acceptation status of an avalanche proof.